Restructure tests.
