+++ tests/btrdb_integration/test_conn.py
